+++ module-api/src/main/kotlin/finn/service/PredictionQueryService.kt
@@ -7,6 +7,7 @@ import finn.queryDto.PredictionDetailQueryDto
 import finn.queryDto.PredictionQueryDto
 import finn.repository.PredictionRepository
 import org.springframework.stereotype.Service
+import java.math.BigDecimal
 import java.util.*
 
 @Service
@@ -36,7 +37,7 @@ class PredictionQueryService(private val predictionRepository: PredictionReposit
         return predictionRepository.getAllTickerRecentScore()
     }
 
-    suspend fun getYesterdayVolatility(tickerId: UUID): Double {
+    suspend fun getYesterdayVolatility(tickerId: UUID): BigDecimal {
         return predictionRepository.getYesterdayVolatilityByTickerId(tickerId)
     }
 }
