+++ module-domain/src/main/kotlin/finn/repository/TickerRepository.kt
@@ -2,6 +2,7 @@ package finn.repository
 
 import finn.entity.query.Ticker
 import finn.queryDto.TickerQueryDto
+import java.math.BigDecimal
 import java.util.*
 
 interface TickerRepository {
@@ -10,7 +11,7 @@ interface TickerRepository {
 
     fun findAll(): List<TickerQueryDto>
 
-    suspend fun getPreviousAtrByTickerId(tickerId: UUID): Double
+    suspend fun getPreviousAtrByTickerId(tickerId: UUID): BigDecimal
 
-    suspend fun updateTodayAtr(tickerId: UUID, todayAtr: Double)
+    suspend fun updateTodayAtr(tickerId: UUID, todayAtr: BigDecimal)
 }
